sql: move tableUpdater to its own file
Release note: None
